Add retries to coder agent start cmd
