+++ agent/doc.go
@@ -1,0 +1,2 @@
+// Package agent is for interacting with p2p server and clients
+package agent
